try to enhance load data
